Dedacted First Commit:
Originally written in early 2008 to help upload stores to cafepress.com.
This is the first version of this functionality.  It's beautiful in it's
simplicity, basically 2 files and it was good enough to get test designs
uploaded and selling.

Requires an account.rb file to provide the login and password for using
the script.